Factor out reducer-helpers
