+++ client/app/lib/helpers/reducer-helpers.js
